+++ scripts/download-sdks.ps1
@@ -47,3 +47,8 @@ foreach ($sdk in $sdks)
 
     gh run download $runId -n "$sdk-sdk" -D $sdkDir
 }
+
+Write-Host "Downloading Crash Reporter binaries ..."
+gh run download $runId -n "CrashReporter-Win64" -D "$outDir/Win64/bin"
+gh run download $runId -n "CrashReporter-Linux" -D "$outDir/Linux/bin"
+gh run download $runId -n "CrashReporter-LinuxArm64" -D "$outDir/LinuxArm64/bin"